% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFun.R
\name{TestFun}
\alias{TestFun}
\title{A funner Hello, World}
\usage{
TestFun(a, b = TRUE)
}
\arguments{
\item{a}{'Numeric: non-zero, positive'}

\item{b}{'Logical: TRUE/FALSE'}
}
\value{
'string'
}
\description{
A funner Hello, World
}
\examples{
TestFun(3, FALSE)
}
\author{
'Sam Hannah'
}
